Restructuring repository
